+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### develop
+
+* Handle "X ounce can" compound unit pattern without leading count (@[MaximusClub](https://github.com/MaximusCub))
+
 ## 2.6.0
 
 * Add `custom_units` keyword argument to ``parse_ingredient`` that allow a dict of custom units to be provided to aid in identifying units. See the [docs](https://ingredient-parser.readthedocs.io/en/develop/tutorials/options.html#custom-units) for more details.